+++ src/main/java/com/jakewharton/rxrelay3/PublishRelay.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Netflix, Inc.
+ * Copyright (c) 2016-present, RxJava Contributors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in
  * compliance with the License. You may obtain a copy of the License at
@@ -11,10 +11,12 @@
  * the License for the specific language governing permissions and limitations under the License.
  */
 
-package com.jakewharton.rxrelay2;
+package com.jakewharton.rxrelay3;
 
-import io.reactivex.Observer;
-import io.reactivex.disposables.Disposable;
+import io.reactivex.rxjava3.core.Observer;
+import io.reactivex.rxjava3.annotations.CheckReturnValue;
+import io.reactivex.rxjava3.annotations.NonNull;
+import io.reactivex.rxjava3.disposables.Disposable;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
@@ -42,14 +44,16 @@ import java.util.concurrent.atomic.AtomicReference;
 public final class PublishRelay<T> extends Relay<T> {
     /** An empty subscribers array to avoid allocating it all the time. */
     @SuppressWarnings("rawtypes")
-    private static final PublishDisposable[] EMPTY = new PublishDisposable[0];
+    static final PublishDisposable[] EMPTY = new PublishDisposable[0];
 
     /** The array of currently subscribed subscribers. */
-    private final AtomicReference<PublishDisposable<T>[]> subscribers;
+    final AtomicReference<PublishDisposable<T>[]> subscribers;
 
     /**
      * Constructs a PublishRelay.
      */
+    @CheckReturnValue
+    @NonNull
     public static <T> PublishRelay<T> create() {
         return new PublishRelay<T>();
     }
@@ -57,13 +61,13 @@ public final class PublishRelay<T> extends Relay<T> {
     /**
      * Constructs a PublishRelay.
      */
-    @SuppressWarnings("unchecked") private PublishRelay() {
+    @SuppressWarnings("unchecked")
+    PublishRelay() {
         subscribers = new AtomicReference<PublishDisposable<T>[]>(EMPTY);
     }
 
-
     @Override
-    public void subscribeActual(Observer<? super T> t) {
+    protected void subscribeActual(Observer<? super T> t) {
         PublishDisposable<T> ps = new PublishDisposable<T>(t, this);
         t.onSubscribe(ps);
         add(ps);
@@ -78,7 +82,7 @@ public final class PublishRelay<T> extends Relay<T> {
      * Adds the given subscriber to the subscribers array atomically.
      * @param ps the subscriber to add
      */
-    private void add(PublishDisposable<T> ps) {
+    void add(PublishDisposable<T> ps) {
         for (;;) {
             PublishDisposable<T>[] a = subscribers.get();
             int n = a.length;
@@ -134,7 +138,7 @@ public final class PublishRelay<T> extends Relay<T> {
     }
 
     @Override
-    public void accept(T value) {
+    public void accept(@NonNull T value) {
         if (value == null) throw new NullPointerException("value == null");
         for (PublishDisposable<T> s : subscribers.get()) {
             s.onNext(value);
@@ -156,7 +160,7 @@ public final class PublishRelay<T> extends Relay<T> {
 
         private static final long serialVersionUID = 3562861878281475070L;
         /** The actual subscriber. */
-        final Observer<? super T> actual;
+        final Observer<? super T> downstream;
         /** The subject state. */
         final PublishRelay<T> parent;
 
@@ -166,13 +170,13 @@ public final class PublishRelay<T> extends Relay<T> {
          * @param parent the parent PublishProcessor
          */
         PublishDisposable(Observer<? super T> actual, PublishRelay<T> parent) {
-            this.actual = actual;
+            this.downstream = actual;
             this.parent = parent;
         }
 
-        void onNext(T t) {
+        public void onNext(T t) {
             if (!get()) {
-                actual.onNext(t);
+                downstream.onNext(t);
             }
         }