+++ src/main/java/com/jakewharton/rxrelay3/ReplayRelay.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2016 Netflix, Inc.
+ * Copyright (c) 2016-present, RxJava Contributors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in
  * compliance with the License. You may obtain a copy of the License at
@@ -11,11 +11,14 @@
  * the License for the specific language governing permissions and limitations under the License.
  */
 
-package com.jakewharton.rxrelay2;
+package com.jakewharton.rxrelay3;
 
-import io.reactivex.Observer;
-import io.reactivex.Scheduler;
-import io.reactivex.disposables.Disposable;
+import io.reactivex.rxjava3.core.Observer;
+import io.reactivex.rxjava3.core.Scheduler;
+import io.reactivex.rxjava3.annotations.CheckReturnValue;
+import io.reactivex.rxjava3.annotations.NonNull;
+import io.reactivex.rxjava3.annotations.Nullable;
+import io.reactivex.rxjava3.disposables.Disposable;
 import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
@@ -62,6 +65,8 @@ public final class ReplayRelay<T> extends Relay<T> {
      * and latency. This can be avoided with the {@link #create(int)} overload which takes an initial capacity
      * parameter and can be tuned to reduce the array reallocation frequency as needed.
      */
+    @CheckReturnValue
+    @NonNull
     public static <T> ReplayRelay<T> create() {
         return new ReplayRelay<T>(new UnboundedReplayBuffer<T>(16));
     }
@@ -78,6 +83,8 @@ public final class ReplayRelay<T> extends Relay<T> {
      * @param capacityHint
      *          the initial buffer capacity
      */
+    @CheckReturnValue
+    @NonNull
     public static <T> ReplayRelay<T> create(int capacityHint) {
         return new ReplayRelay<T>(new UnboundedReplayBuffer<T>(capacityHint));
     }
@@ -99,6 +106,8 @@ public final class ReplayRelay<T> extends Relay<T> {
      * @param maxSize
      *          the maximum number of buffered items
      */
+    @CheckReturnValue
+    @NonNull
     public static <T> ReplayRelay<T> createWithSize(int maxSize) {
         return new ReplayRelay<T>(new SizeBoundReplayBuffer<T>(maxSize));
     }
@@ -145,6 +154,8 @@ public final class ReplayRelay<T> extends Relay<T> {
      * @param scheduler
      *          the {@link Scheduler} that provides the current time
      */
+    @CheckReturnValue
+    @NonNull
     public static <T> ReplayRelay<T> createWithTime(long maxAge, TimeUnit unit, Scheduler scheduler) {
         return new ReplayRelay<T>(new SizeAndTimeBoundReplayBuffer<T>(Integer.MAX_VALUE, maxAge, unit, scheduler));
     }
@@ -180,6 +191,8 @@ public final class ReplayRelay<T> extends Relay<T> {
      * @param scheduler
      *          the {@link Scheduler} that provides the current time
      */
+    @CheckReturnValue
+    @NonNull
     public static <T> ReplayRelay<T> createWithTimeAndSize(long maxAge, TimeUnit unit, Scheduler scheduler, int maxSize) {
         return new ReplayRelay<T>(new SizeAndTimeBoundReplayBuffer<T>(maxSize, maxAge, unit, scheduler));
     }
@@ -210,7 +223,7 @@ public final class ReplayRelay<T> extends Relay<T> {
     }
 
     @Override
-    public void accept(T value) {
+    public void accept(@NonNull T value) {
         if (value == null) throw new NullPointerException("value == null");
 
         ReplayBuffer<T> b = buffer;
@@ -234,8 +247,26 @@ public final class ReplayRelay<T> extends Relay<T> {
      * Returns a single value the Relay currently has or null if no such value exists.
      * <p>The method is thread-safe.
      */
+    @Nullable
     public T getValue() {
         return buffer.getValue();
+    }
+
+    /**
+     * Makes sure the item cached by the head node in a bounded
+     * ReplayRelay is released (as it is never part of a replay).
+     * <p>
+     * By default, live bounded buffers will remember one item before
+     * the currently receivable one to ensure subscribers can always
+     * receive a continuous sequence of items. A terminated ReplaySubject
+     * automatically releases this inaccessible item.
+     * <p>
+     * The method must be called sequentially, similar to the standard
+     * {@code onXXX} methods.
+     * @since 2.1
+     */
+    public void cleanupBuffer() {
+        buffer.trimHead();
     }
 
     /** An empty array to avoid allocation in getValues(). */
@@ -338,15 +369,18 @@ public final class ReplayRelay<T> extends Relay<T> {
 
         int size();
 
+        @Nullable
         T getValue();
 
         T[] getValues(T[] array);
+
+        void trimHead();
     }
 
     static final class ReplayDisposable<T> extends AtomicInteger implements Disposable {
 
         private static final long serialVersionUID = 466549804534799122L;
-        final Observer<? super T> actual;
+        final Observer<? super T> downstream;
         final ReplayRelay<T> state;
 
         Object index;
@@ -354,7 +388,7 @@ public final class ReplayRelay<T> extends Relay<T> {
         volatile boolean cancelled;
 
         ReplayDisposable(Observer<? super T> actual, ReplayRelay<T> state) {
-            this.actual = actual;
+            this.downstream = actual;
             this.state = state;
         }
 
@@ -383,7 +417,7 @@ public final class ReplayRelay<T> extends Relay<T> {
         volatile int size;
 
         UnboundedReplayBuffer(int capacityHint) {
-            if (capacityHint <= 0) throw new IllegalArgumentException("capacityHint <= 0");
+            if (capacityHint <= 0) throw new IllegalArgumentException("capacityHint > 0 required but it was " + capacityHint);
             this.buffer = new ArrayList<T>(capacityHint);
         }
 
@@ -394,6 +428,12 @@ public final class ReplayRelay<T> extends Relay<T> {
         }
 
         @Override
+        public void trimHead() {
+            // no-op in this type of buffer
+        }
+
+        @Override
+        @Nullable
         @SuppressWarnings("unchecked")
         public T getValue() {
             int s = size;
@@ -437,7 +477,7 @@ public final class ReplayRelay<T> extends Relay<T> {
 
             int missed = 1;
             final List<T> b = buffer;
-            final Observer<? super T> a = rs.actual;
+            final Observer<? super T> a = rs.downstream;
 
             Integer indexObject = (Integer)rs.index;
             int index;
@@ -466,7 +506,7 @@ public final class ReplayRelay<T> extends Relay<T> {
 
                     T o = b.get(index);
 
-                    a.onNext((T)o);
+                    a.onNext(o);
                     index++;
                 }
 
@@ -485,8 +525,7 @@ public final class ReplayRelay<T> extends Relay<T> {
 
         @Override
         public int size() {
-            int s = size;
-            return s != 0 ? s : 0;
+            return size;
         }
     }
 
@@ -557,7 +596,22 @@ public final class ReplayRelay<T> extends Relay<T> {
             trim();
         }
 
+        /**
+         * Replace a non-empty head node with an empty one to
+         * allow the GC of the inaccessible old value.
+         */
         @Override
+        public void trimHead() {
+            Node<T> h = head;
+            if (h.value != null) {
+                Node<T> n = new Node<T>(null);
+                n.lazySet(h.get());
+                head = n;
+            }
+        }
+
+        @Override
+        @Nullable
         @SuppressWarnings("unchecked")
         public T getValue() {
             Node<T> h = head;
@@ -591,7 +645,7 @@ public final class ReplayRelay<T> extends Relay<T> {
                 int i = 0;
                 while (i != s) {
                     Node<T> next = h.get();
-                    array[i] = (T)next.value;
+                    array[i] = next.value;
                     i++;
                     h = next;
                 }
@@ -611,7 +665,7 @@ public final class ReplayRelay<T> extends Relay<T> {
             }
 
             int missed = 1;
-            final Observer<? super T> a = rs.actual;
+            final Observer<? super T> a = rs.downstream;
 
             Node<T> index = (Node<T>)rs.index;
             if (index == null) {
@@ -632,9 +686,7 @@ public final class ReplayRelay<T> extends Relay<T> {
                         break;
                     }
 
-                    T o = n.value;
-
-                    a.onNext((T)o);
+                    a.onNext(n.value);
 
                     index = n;
                 }
@@ -714,6 +766,10 @@ public final class ReplayRelay<T> extends Relay<T> {
             TimedNode<T> h = head;
 
             for (;;) {
+                if (size <= 1) {
+                    head = h;
+                    break;
+                }
                 TimedNode<T> next = h.get();
                 if (next == null) {
                     head = h;
@@ -726,6 +782,7 @@ public final class ReplayRelay<T> extends Relay<T> {
                 }
 
                 h = next;
+                size--;
             }
 
         }
@@ -742,7 +799,22 @@ public final class ReplayRelay<T> extends Relay<T> {
             trim();
         }
 
+        /**
+         * Replace a non-empty head node with an empty one to
+         * allow the GC of the inaccessible old value.
+         */
         @Override
+        public void trimHead() {
+            TimedNode<T> h = head;
+            if (h.value != null) {
+                TimedNode<T> n = new TimedNode<T>(null, 0);
+                n.lazySet(h.get());
+                head = n;
+            }
+        }
+
+        @Override
+        @Nullable
         @SuppressWarnings("unchecked")
         public T getValue() {
             TimedNode<T> h = head;
@@ -755,14 +827,35 @@ public final class ReplayRelay<T> extends Relay<T> {
                 h = next;
             }
 
+            long limit = scheduler.now(unit) - maxAge;
+            if (h.time < limit) {
+                return null;
+            }
+
             return h.value;
+        }
+
+        TimedNode<T> getHead() {
+            TimedNode<T> index = head;
+            // skip old entries
+            long limit = scheduler.now(unit) - maxAge;
+            TimedNode<T> next = index.get();
+            while (next != null) {
+                long ts = next.time;
+                if (ts > limit) {
+                    break;
+                }
+                index = next;
+                next = index.get();
+            }
+            return index;
         }
 
         @Override
         @SuppressWarnings("unchecked")
         public T[] getValues(T[] array) {
-            TimedNode<T> h = head;
-            int s = size();
+            TimedNode<T> h = getHead();
+            int s = size(h);
 
             if (s == 0) {
                 if (array.length != 0) {
@@ -796,22 +889,11 @@ public final class ReplayRelay<T> extends Relay<T> {
             }
 
             int missed = 1;
-            final Observer<? super T> a = rs.actual;
+            final Observer<? super T> a = rs.downstream;
 
             TimedNode<T> index = (TimedNode<T>)rs.index;
             if (index == null) {
-                index = head;
-                // skip old entries
-                long limit = scheduler.now(unit) - maxAge;
-                TimedNode<T> next = index.get();
-                while (next != null) {
-                    long ts = next.time;
-                    if (ts > limit) {
-                        break;
-                    }
-                    index = next;
-                    next = index.get();
-                }
+                index = getHead();
             }
 
             for (;;) {
@@ -855,8 +937,11 @@ public final class ReplayRelay<T> extends Relay<T> {
 
         @Override
         public int size() {
+            return size(getHead());
+        }
+
+        int size(TimedNode<T> h) {
             int s = 0;
-            TimedNode<T> h = head;
             while (s != Integer.MAX_VALUE) {
                 TimedNode<T> next = h.get();
                 if (next == null) {