+++ src/main/java/com/jakewharton/rxrelay3/SerializedRelay.java
@@ -11,9 +11,10 @@
  * the License for the specific language governing permissions and limitations under the License.
  */
 
-package com.jakewharton.rxrelay2;
+package com.jakewharton.rxrelay3;
 
-import io.reactivex.Observer;
+import io.reactivex.rxjava3.annotations.NonNull;
+import io.reactivex.rxjava3.core.Observer;
 
 /**
  * Serializes calls to the Consumer methods.
@@ -41,7 +42,7 @@ import io.reactivex.Observer;
 
 
     @Override
-    public void accept(T value) {
+    public void accept(@NonNull T value) {
         synchronized (this) {
             if (emitting) {
                 AppendOnlyLinkedArrayList<T> q = queue;